Change double variables to long double

+++ core/src/BootstrapTree.hpp
@@ -24,11 +24,11 @@ namespace models {
       : Base(std::move(root),  std::move(training_spec), training_data) {
     }
 
-    double error_rate(const stats::DataSpec<T, R> &data) const override {
+    long double error_rate(const stats::DataSpec<T, R> &data) const override {
       return Tree<T, R, stats::BootstrapDataSpec<T, R> >::error_rate(data);
     }
 
-    double error_rate() const {
+    long double error_rate() const {
       return error_rate(this->training_data->get_oob());
     }
 
@@ -47,7 +47,7 @@ namespace models {
 
       math::DVector<T> importance = Base::variable_importance(importance_kind);
 
-      double factor = 1.0;
+      long double factor = 1.0;
 
       if (importance_kind == VariableImportanceKind::PROJECTOR_ADJUSTED) {
         factor = (1.0 - error_rate());
@@ -60,14 +60,14 @@ namespace models {
       math::DVector<T> permutation_variable_importance() const {
         const stats::DataSpec<T, R> oob = this->training_data->get_oob();
         const stats::DataColumn<R> oob_predictions = Base::predict(oob.x);
-        const double oob_accuracy = stats::accuracy(oob_predictions, oob.y);
+        const long double oob_accuracy = stats::accuracy(oob_predictions, oob.y);
 
         math::DVector<T> importance = math::DVector<T>(oob.x.cols());
 
         for (int j = 0; j < oob.x.cols(); j++) {
           const stats::Data<T> nonsense_data = stats::shuffle_column(oob.x, j);
           const stats::DataColumn<R> nonsense_predictions = Base::predict(nonsense_data);
-          const double nonsense_accuracy = stats::accuracy(nonsense_predictions, oob.y);
+          const long double nonsense_accuracy = stats::accuracy(nonsense_predictions, oob.y);
 
           importance(j) = oob_accuracy - nonsense_accuracy;
         }
