+++ core/src/Projector.hpp
@@ -14,12 +14,12 @@ namespace models::pp {
     using Size = Eigen::Index;
 
     const math::DVector<T> vector;
-    const double index;
+    const long double index;
 
-    explicit Projector(const math::DVector<T> vector, const double index = 0) : vector(vector), index(index) {
+    explicit Projector(const math::DVector<T> vector, const long double index = 0) : vector(vector), index(index) {
     }
 
-    Projector(std::initializer_list<T> init_list, const double index = 0)
+    Projector(std::initializer_list<T> init_list, const long double index = 0)
       : vector(math::DVector<T>::Map(init_list.begin(), init_list.size())), index(index) {
     }
